refactor(gateway): extract source reply finalization
